More splitting up and doc improvements.
