+++ doc/sphinx/concepts.rst
@@ -14,6 +14,10 @@ Common concepts
 
 .. cpp:concept:: template <typename T> mppp::CppInteroperable
 
+   Interoperable fundamental C++ types.
+
+   *#include <mp++/concepts.hpp>*
+
    This concept is satisfied by any C++ fundamental type with which the multiprecision classes (such as :cpp:class:`~mppp::integer`)
    can interoperate. The full list of types satisfying this concept is:
